refactor: move record encoding into storage
