+++ storage/src/record/encoding.rs
@@ -1,0 +1,562 @@
+use std::num::NonZeroU8;
+
+#[cfg(test)]
+use bytes::BytesMut;
+use bytes::{Buf, BufMut, Bytes};
+use s2_common::record::{
+    CommandOp, CommandPayloadError, CommandRecord, EnvelopeRecord, Header, HeaderValidationError,
+    RecordPartsError,
+};
+
+#[derive(Debug, Clone, PartialEq, Eq, thiserror::Error)]
+pub enum RecordDecodeError {
+    #[error("truncated: {0}")]
+    Truncated(&'static str),
+    #[error("invalid value [{0}]: {1}")]
+    InvalidValue(&'static str, &'static str),
+}
+
+pub(crate) trait Encodable {
+    #[cfg(test)]
+    fn to_bytes(&self) -> Bytes {
+        let expected_size = self.encoded_size();
+        let mut buf = BytesMut::with_capacity(expected_size);
+        self.encode_into(&mut buf);
+        assert_eq!(buf.len(), expected_size, "no reallocation");
+        buf.freeze()
+    }
+
+    fn encoded_size(&self) -> usize;
+
+    fn encode_into(&self, buf: &mut impl BufMut);
+}
+
+const COMMAND_ORDINAL_FENCE: u8 = 0;
+const COMMAND_ORDINAL_TRIM: u8 = 1;
+
+fn command_op_ordinal(op: CommandOp) -> u8 {
+    match op {
+        CommandOp::Fence => COMMAND_ORDINAL_FENCE,
+        CommandOp::Trim => COMMAND_ORDINAL_TRIM,
+    }
+}
+
+fn command_op_from_ordinal(ordinal: u8) -> Option<CommandOp> {
+    match ordinal {
+        COMMAND_ORDINAL_FENCE => Some(CommandOp::Fence),
+        COMMAND_ORDINAL_TRIM => Some(CommandOp::Trim),
+        _ => None,
+    }
+}
+
+impl From<CommandPayloadError> for RecordDecodeError {
+    fn from(e: CommandPayloadError) -> Self {
+        match e {
+            CommandPayloadError::InvalidUtf8(_) => {
+                RecordDecodeError::InvalidValue("CommandPayload", "fencing token not valid utf8")
+            }
+            CommandPayloadError::FencingTokenTooLong(_) => {
+                RecordDecodeError::InvalidValue("CommandPayload", "fencing token too long")
+            }
+            CommandPayloadError::TrimPointSize(_) => {
+                RecordDecodeError::InvalidValue("CommandPayload", "trim point size")
+            }
+        }
+    }
+}
+
+impl Encodable for CommandRecord {
+    fn encoded_size(&self) -> usize {
+        1 + match self {
+            CommandRecord::Fence(token) => token.len(),
+            CommandRecord::Trim(trim_point) => size_of_val(trim_point),
+        }
+    }
+
+    fn encode_into(&self, buf: &mut impl BufMut) {
+        buf.put_u8(command_op_ordinal(self.op()));
+        match self {
+            CommandRecord::Fence(token) => {
+                buf.put_slice(token.as_bytes());
+            }
+            CommandRecord::Trim(trim_point) => {
+                buf.put_u64(*trim_point);
+            }
+        }
+    }
+}
+
+pub(super) fn decode_command_record(record: &[u8]) -> Result<CommandRecord, RecordDecodeError> {
+    if record.is_empty() {
+        return Err(RecordDecodeError::Truncated("CommandOrdinal"));
+    }
+    let op = command_op_from_ordinal(record[0])
+        .ok_or(RecordDecodeError::InvalidValue("CommandOrdinal", "unknown"))?;
+    CommandRecord::try_from_parts(op, &record[1..]).map_err(Into::into)
+}
+
+const EMPTY_HEADER_FLAG: HeaderFlag = HeaderFlag {
+    num_headers_length_bytes: 0,
+    name_length_bytes: NonZeroU8::new(1).unwrap(),
+    value_length_bytes: NonZeroU8::new(1).unwrap(),
+};
+
+#[derive(Debug, PartialEq, Eq, Clone, Copy)]
+struct HeaderFlag {
+    num_headers_length_bytes: u8,
+    name_length_bytes: NonZeroU8,
+    value_length_bytes: NonZeroU8,
+}
+
+impl From<HeaderFlag> for u8 {
+    fn from(value: HeaderFlag) -> Self {
+        (value.num_headers_length_bytes << 4)
+            | ((value.name_length_bytes.get() - 1) << 2)
+            | (value.value_length_bytes.get() - 1)
+    }
+}
+
+impl TryFrom<u8> for HeaderFlag {
+    type Error = &'static str;
+
+    fn try_from(value: u8) -> Result<Self, Self::Error> {
+        if (value & (0b11u8 << 6)) != 0u8 {
+            return Err("reserved bit set");
+        }
+        Ok(Self {
+            num_headers_length_bytes: (0b110000 & value) >> 4,
+            name_length_bytes: NonZeroU8::new(((0b1100 & value) >> 2) + 1).unwrap(),
+            value_length_bytes: NonZeroU8::new((0b11 & value) + 1).unwrap(),
+        })
+    }
+}
+
+const EMPTY_HEADERS_ENCODING_INFO: EncodingInfo = EncodingInfo {
+    headers_total_bytes: 0,
+    flag: EMPTY_HEADER_FLAG,
+};
+
+#[derive(Debug, PartialEq, Eq, Clone, Copy)]
+struct EncodingInfo {
+    headers_total_bytes: usize,
+    flag: HeaderFlag,
+}
+
+impl TryFrom<&[Header]> for EncodingInfo {
+    type Error = HeaderValidationError;
+
+    fn try_from(headers: &[Header]) -> Result<Self, Self::Error> {
+        fn size_bytes_headers_len(elems: u64) -> Result<u8, HeaderValidationError> {
+            let size = 8 - elems.leading_zeros() / 8;
+            if size > 3 {
+                Err(HeaderValidationError::TooMany)
+            } else {
+                Ok(size as u8)
+            }
+        }
+
+        fn size_bytes_name_value_len(elems: u64) -> Result<NonZeroU8, HeaderValidationError> {
+            if elems == 0 {
+                return Ok(NonZeroU8::new(1u8).unwrap());
+            }
+            let size = 8 - (elems.leading_zeros() / 8);
+            if size > 4 {
+                Err(HeaderValidationError::TooLong)
+            } else {
+                Ok(NonZeroU8::new(size as u8).unwrap())
+            }
+        }
+
+        if headers.is_empty() {
+            return Ok(EMPTY_HEADERS_ENCODING_INFO);
+        }
+
+        let (headers_total_bytes, name_max, value_max) = headers.iter().try_fold(
+            (0usize, 0usize, 0usize),
+            |(size_bytes_acc, name_max, value_max), Header { name, value }| {
+                if name.is_empty() {
+                    return Err(HeaderValidationError::NameEmpty);
+                }
+                let name_len = name.len();
+                let value_len = value.len();
+                Ok((
+                    size_bytes_acc + name_len + value_len,
+                    name_max.max(name_len),
+                    value_max.max(value_len),
+                ))
+            },
+        )?;
+
+        let num_headers_length_bytes = size_bytes_headers_len(headers.len() as u64)?;
+        let name_length_bytes = size_bytes_name_value_len(name_max as u64)?;
+        let value_length_bytes = size_bytes_name_value_len(value_max as u64)?;
+
+        Ok(Self {
+            headers_total_bytes,
+            flag: HeaderFlag {
+                num_headers_length_bytes,
+                name_length_bytes,
+                value_length_bytes,
+            },
+        })
+    }
+}
+
+impl Encodable for EnvelopeRecord {
+    fn encoded_size(&self) -> usize {
+        let encoding_info = EncodingInfo::try_from(self.headers())
+            .expect("envelope record headers should be validated");
+        1 + encoding_info.flag.num_headers_length_bytes as usize
+            + self.headers().len()
+                * (encoding_info.flag.name_length_bytes.get() as usize
+                    + encoding_info.flag.value_length_bytes.get() as usize)
+            + encoding_info.headers_total_bytes
+            + self.body().len()
+    }
+
+    fn encode_into(&self, buf: &mut impl BufMut) {
+        let encoding_info = EncodingInfo::try_from(self.headers())
+            .expect("envelope record headers should be validated");
+        buf.put_u8(encoding_info.flag.into());
+        buf.put_uint(
+            self.headers().len() as u64,
+            encoding_info.flag.num_headers_length_bytes as usize,
+        );
+        for Header { name, value } in self.headers() {
+            buf.put_uint(
+                name.len() as u64,
+                encoding_info.flag.name_length_bytes.get() as usize,
+            );
+            buf.put_slice(name);
+            buf.put_uint(
+                value.len() as u64,
+                encoding_info.flag.value_length_bytes.get() as usize,
+            );
+            buf.put_slice(value);
+        }
+        buf.put_slice(self.body());
+    }
+}
+
+pub(super) fn decode_envelope_record(mut buf: Bytes) -> Result<EnvelopeRecord, RecordDecodeError> {
+    if buf.is_empty() {
+        return Err(RecordDecodeError::InvalidValue("HeaderFlag", "missing"));
+    }
+
+    let flag: HeaderFlag = buf
+        .get_u8()
+        .try_into()
+        .map_err(|info| RecordDecodeError::InvalidValue("HeaderFlag", info))?;
+    if flag.num_headers_length_bytes == 0 {
+        return EnvelopeRecord::try_from_parts(vec![], buf).map_err(record_parts_decode_error);
+    }
+
+    let num_headers = buf
+        .try_get_uint(flag.num_headers_length_bytes as usize)
+        .map_err(|_| RecordDecodeError::Truncated("NumHeaders"))?;
+    let num_headers = usize::try_from(num_headers)
+        .map_err(|_| RecordDecodeError::InvalidValue("NumHeaders", "too many"))?;
+
+    let mut headers: Vec<Header> = Vec::with_capacity(num_headers);
+    for _ in 0..num_headers {
+        let name_len =
+            buf.try_get_uint(flag.name_length_bytes.get() as usize)
+                .map_err(|_| RecordDecodeError::Truncated("HeaderNameLen"))? as usize;
+        if name_len == 0 {
+            return Err(RecordDecodeError::InvalidValue("HeaderName", "empty"));
+        }
+        if buf.remaining() < name_len {
+            return Err(RecordDecodeError::Truncated("HeaderName"));
+        }
+        let name = buf.split_to(name_len);
+
+        let value_len =
+            buf.try_get_uint(flag.value_length_bytes.get() as usize)
+                .map_err(|_| RecordDecodeError::Truncated("HeaderValueLen"))? as usize;
+        if buf.remaining() < value_len {
+            return Err(RecordDecodeError::Truncated("HeaderValue"));
+        }
+        let value = buf.split_to(value_len);
+
+        headers.push(Header { name, value })
+    }
+
+    EnvelopeRecord::try_from_parts(headers, buf).map_err(record_parts_decode_error)
+}
+
+fn record_parts_decode_error(error: RecordPartsError) -> RecordDecodeError {
+    match error {
+        RecordPartsError::Header(HeaderValidationError::NameEmpty) => {
+            RecordDecodeError::InvalidValue("HeaderName", "empty")
+        }
+        RecordPartsError::Header(HeaderValidationError::TooMany) => {
+            RecordDecodeError::InvalidValue("NumHeaders", "too many")
+        }
+        RecordPartsError::Header(HeaderValidationError::TooLong) => {
+            RecordDecodeError::InvalidValue("Header", "too long")
+        }
+        RecordPartsError::UnknownCommand | RecordPartsError::CommandPayload(_, _) => {
+            RecordDecodeError::InvalidValue("EnvelopeRecord", "unexpected command record")
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use bytes::{BufMut, Bytes, BytesMut};
+    use rstest::rstest;
+    use s2_common::record::{FencingToken, FencingTokenTooLongError, SeqNum};
+
+    use super::*;
+
+    fn roundtrip_command(cmd: CommandRecord, expected_len: usize) {
+        assert_eq!(cmd.encoded_size(), expected_len);
+        let encoded = cmd.to_bytes();
+        assert_eq!(encoded.len(), expected_len);
+        assert_eq!(decode_command_record(encoded.as_ref()), Ok(cmd));
+    }
+
+    #[rstest]
+    #[case::empty("")]
+    #[case::arbit("arbitrary")]
+    #[case::full("0123456789012345")]
+    fn command_fence_roundtrip(#[case] token: &str) {
+        let cmd = CommandRecord::Fence(token.parse::<FencingToken>().unwrap());
+        roundtrip_command(cmd, 1 + token.len());
+    }
+
+    #[rstest]
+    #[case::zero(0)]
+    #[case::large(SeqNum::MAX)]
+    fn command_trim_roundtrip(#[case] trim_point: SeqNum) {
+        roundtrip_command(CommandRecord::Trim(trim_point), 1 + size_of::<SeqNum>());
+    }
+
+    #[test]
+    fn decode_invalid_command() {
+        let try_convert = |raw: &[u8]| decode_command_record(raw);
+        assert_eq!(
+            try_convert(&[]),
+            Err(RecordDecodeError::Truncated("CommandOrdinal"))
+        );
+        assert_eq!(
+            try_convert(&[0xff]),
+            Err(RecordDecodeError::InvalidValue("CommandOrdinal", "unknown"))
+        );
+        assert_eq!(
+            try_convert(&[command_op_ordinal(CommandOp::Fence), 0xff, 0xff]),
+            Err(RecordDecodeError::InvalidValue(
+                "CommandPayload",
+                "fencing token not valid utf8"
+            ))
+        );
+        assert_eq!(
+            try_convert(&[
+                command_op_ordinal(CommandOp::Fence),
+                b'0',
+                b'1',
+                b'2',
+                b'3',
+                b'4',
+                b'5',
+                b'6',
+                b'7',
+                b'8',
+                b'9',
+                b'0',
+                b'1',
+                b'2',
+                b'3',
+                b'4',
+                b'5',
+                b'6',
+                b'7',
+                b'8',
+                b'9',
+                b'0',
+                b'1',
+                b'2',
+                b'3',
+                b'4',
+                b'5',
+                b'6',
+                b'7',
+                b'8',
+                b'9',
+                b'0',
+                b'1',
+                b'2',
+                b'3',
+                b'4',
+                b'5',
+                b'6',
+                b'7',
+                b'8',
+                b'9',
+            ]),
+            Err(CommandPayloadError::FencingTokenTooLong(FencingTokenTooLongError(40)).into())
+        );
+        assert_eq!(
+            try_convert(&[command_op_ordinal(CommandOp::Trim), 0xff]),
+            Err(CommandPayloadError::TrimPointSize(1).into())
+        );
+    }
+
+    fn roundtrip_envelope_parts(headers: Vec<Header>, body: Bytes) {
+        let encoded: Bytes = EnvelopeRecord::try_from_parts(headers.clone(), body.clone())
+            .unwrap()
+            .to_bytes();
+        let decoded = decode_envelope_record(encoded).unwrap();
+        assert_eq!(decoded.headers(), headers);
+        assert_eq!(decoded.body(), &body);
+    }
+
+    #[test]
+    fn envelope_framed_with_headers() {
+        roundtrip_envelope_parts(
+            vec![
+                Header {
+                    name: Bytes::from("key_1"),
+                    value: Bytes::from("val_1"),
+                },
+                Header {
+                    name: Bytes::from("key_2"),
+                    value: Bytes::from("val_2"),
+                },
+                Header {
+                    name: Bytes::from("key_3"),
+                    value: Bytes::from("val_3"),
+                },
+                Header {
+                    name: Bytes::from("key_4"),
+                    value: Bytes::from("val_4"),
+                },
+            ],
+            Bytes::from("hello"),
+        );
+    }
+
+    #[test]
+    fn envelope_framed_no_headers() {
+        roundtrip_envelope_parts(vec![], Bytes::from("hello"));
+    }
+
+    #[test]
+    fn envelope_decode_rejects_empty_header_name() {
+        let mut encoded = BytesMut::new();
+        encoded.put_u8(
+            HeaderFlag {
+                num_headers_length_bytes: 1,
+                name_length_bytes: NonZeroU8::new(1).unwrap(),
+                value_length_bytes: NonZeroU8::new(1).unwrap(),
+            }
+            .into(),
+        );
+        encoded.put_u8(1);
+        encoded.put_u8(0);
+        encoded.put_u8(5);
+        encoded.put_slice(b"value");
+        encoded.put_slice(b"body");
+
+        assert_eq!(
+            decode_envelope_record(encoded.freeze()),
+            Err(RecordDecodeError::InvalidValue("HeaderName", "empty"))
+        );
+    }
+
+    #[test]
+    fn envelope_framed_duplicate_keys() {
+        roundtrip_envelope_parts(
+            vec![
+                Header {
+                    name: Bytes::from("b"),
+                    value: Bytes::from("val_1"),
+                },
+                Header {
+                    name: Bytes::from("b"),
+                    value: Bytes::from("val_2"),
+                },
+                Header {
+                    name: Bytes::from("a"),
+                    value: Bytes::from("val_3"),
+                },
+            ],
+            Bytes::from("hello"),
+        );
+    }
+
+    #[test]
+    fn flag_ex1() {
+        assert_eq!(
+            Ok(HeaderFlag {
+                num_headers_length_bytes: 2,
+                name_length_bytes: NonZeroU8::new(1).unwrap(),
+                value_length_bytes: NonZeroU8::new(1).unwrap(),
+            }),
+            0b00100000.try_into()
+        );
+
+        let u8_repr: u8 = HeaderFlag {
+            num_headers_length_bytes: 2,
+            name_length_bytes: NonZeroU8::new(1).unwrap(),
+            value_length_bytes: NonZeroU8::new(1).unwrap(),
+        }
+        .into();
+        assert_eq!(u8_repr, 0b00100000);
+    }
+
+    #[test]
+    fn flag_ex2() {
+        assert_eq!(
+            Ok(HeaderFlag {
+                num_headers_length_bytes: 1,
+                name_length_bytes: NonZeroU8::new(1).unwrap(),
+                value_length_bytes: NonZeroU8::new(1).unwrap(),
+            }),
+            0b00010000.try_into()
+        );
+
+        let u8_repr: u8 = HeaderFlag {
+            num_headers_length_bytes: 1,
+            name_length_bytes: NonZeroU8::new(1).unwrap(),
+            value_length_bytes: NonZeroU8::new(1).unwrap(),
+        }
+        .into();
+        assert_eq!(u8_repr, 0b00010000);
+    }
+
+    #[test]
+    fn empty_envelope_size() {
+        assert_eq!(
+            1,
+            EnvelopeRecord::try_from_parts(vec![], Bytes::new())
+                .unwrap()
+                .to_bytes()
+                .len()
+        );
+    }
+
+    #[test]
+    fn truncated_envelope_returns_error() {
+        let record = EnvelopeRecord::try_from_parts(
+            vec![Header {
+                name: Bytes::from("key"),
+                value: Bytes::from("value"),
+            }],
+            Bytes::new(),
+        )
+        .unwrap();
+        let encoded = record.to_bytes();
+
+        for len in 1..encoded.len() {
+            let truncated = encoded.slice(..len);
+            assert!(
+                matches!(
+                    decode_envelope_record(truncated),
+                    Err(RecordDecodeError::Truncated(_))
+                ),
+                "expected Truncated error for len {len}"
+            );
+        }
+    }
+}
